# apps/backend/design_docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ACMG-PS3 Intelligent Parsing API
  description: |-
    Versioned REST contract for the FastAPI backend that powers the ACMG-PS3
    multilingual document evidence extraction platform. The specification mirrors
    the current `/api/v1` implementation (FastAPI routers in
    `src/presentation/controllers`) so that the generated Swagger UI stays in sync
    with the running service.

    All endpoints are asynchronous-friendly. Uploads immediately return a task
    identifier, while task state can be polled via REST or streamed through the
    `/ws/task/{task_id}/progress` WebSocket.
  version: 2.1.0
  contact:
    name: ACMG-PS3 Engineering
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: FastAPI root (dev)
  - url: http://localhost:8000/api/v1
    description: Versioned REST prefix
...
paths:
  /api/v1/pdf/upload:
    post:
      tags: [PDF Parsing]
      summary: Upload PDF via JSON payload
      description: >
        Accepts base64-encoded PDF bytes or PMID/DOI metadata in a JSON payload
        and creates an asynchronous parsing task. The endpoint mirrors
        `PDFParseController.upload_pdf`.
      operationId: pdfUploadJson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PDFUploadRequest'
      responses:
        '202':
          description: Upload accepted; parsing scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreatedResponse'
        '400':
          description: Validation error (invalid base64, missing filename, oversized file).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error while creating the task.
  /api/v1/pdf/upload/form:
    post:
      tags: [PDF Parsing]
      summary: Upload PDF via multipart/form-data
      operationId: pdfUploadForm
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF binary (max 50 MB by default).
                priority:
                  type: integer
                  minimum: 0
                  maximum: 10
                  default: 0
                  description: Optional queue priority override.
      responses:
        '202':
          description: Upload accepted; parsing scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreatedResponse'
        '400':
          description: Invalid file or payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error while creating the task.
  /api/v1/pdf/fetch-by-pmid:
    post:
      tags: [PDF Parsing]
      summary: Create parsing task from a PubMed ID
      operationId: pdfFetchByPMID
      parameters:
        - name: pmid
          in: query
          required: true
          schema:
            type: string
            example: "12345678"
          description: PubMed identifier (1-8 digits).
        - name: priority
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 10
            default: 0
          description: Optional queue priority override.
      responses:
        '202':
          description: Fetch request accepted and task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreatedResponse'
        '400':
          description: Invalid PMID format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
  /api/v1/pdf/fetch-by-doi:
    post:
      tags: [PDF Parsing]
      summary: Create parsing task from a DOI
      operationId: pdfFetchByDOI
      parameters:
        - name: doi
          in: query
          required: true
          schema:
            type: string
            example: "10.1038/s41467-024-00001-0"
          description: Digital Object Identifier for the publication.
        - name: priority
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 10
            default: 0
          description: Optional queue priority override.
      responses:
        '202':
          description: Fetch request accepted and task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreatedResponse'
        '400':
          description: Invalid DOI format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
  /api/v1/tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/TaskIdPath'
    get:
      tags: [Task Management]
      summary: Get parsing task status
      description: >
        Returns the persisted status for a parsing task stored in PostgreSQL,
        including evidence summaries if available.
      operationId: parsingTaskStatus
      responses:
        '200':
          description: Current task status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsingTaskStatusResponse'
        '404':
          description: Task does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Task Management]
      summary: Cancel parsing task
      description: Marks an in-flight parsing task as cancelled.
      operationId: parsingTaskCancel
      responses:
        '200':
          description: Cancellation acknowledged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskActionResponse'
        '404':
          description: Task not found or already terminal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/{task_id}/progress:
    parameters:
      - $ref: '#/components/parameters/TaskIdPath'
    get:
      tags: [Task Management]
      summary: Get parsing task progress snapshot
      description: >
        Lightweight polling endpoint that surfaces progress percentage and stage.
        Front-ends can fall back to this endpoint when WebSocket upgrades are not
        available.
      operationId: parsingTaskProgress
      responses:
        '200':
          description: Latest progress snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProgressResponse'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/{task_id}/retry:
    parameters:
      - $ref: '#/components/parameters/TaskIdPath'
    post:
      tags: [Task Management]
      summary: Retry failed parsing task
      description: >
        Resets a task in `failed` status back to `pending`. Only available for
        failure states.
      operationId: parsingTaskRetry
      responses:
        '200':
          description: Retry accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskActionResponse'
        '404':
          description: Task not found or not eligible for retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ws/task/{task_id}/progress:
    parameters:
      - $ref: '#/components/parameters/TaskIdPath'
    get:
      tags: [Realtime Progress]
      summary: Upgrade to WebSocket for live progress updates
      description: >
        Upgrades the HTTP connection to a WebSocket (managed by
        `src/presentation/websocket/progress_handler.ProgressHandler`) and
        streams JSON progress events published by Redis.
      operationId: progressWebSocket
      responses:
        '101':
          description: WebSocket handshake complete.
        '404':
          description: No handler available (route missing in deployment).
components:
  parameters:
    TaskIdPath:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        description: UUID of the parsing task.
  schemas:
    PDFUploadRequest:
      type: object
      description: Payload accepted by `/api/v1/pdf/upload`.
      properties:
        file_content:
          type: string
          description: Base64-encoded PDF content.
          example: JVBERi0xLjQKJcfs...
        filename:
          type: string
          description: Original filename (required when `source=file`).
          example: acmg_evidence_study.pdf
        pmid:
          type: string
          description: PubMed ID (used when `source=pmid`).
          example: "12345678"
        doi:
          type: string
          description: DOI (used when `source=doi`).
          example: 10.1038/s41467-024-00001-0
        source:
          type: string
          description: Upload source selector.
          enum: [file, pmid, doi]
          default: file
        priority:
          type: integer
          minimum: 0
          maximum: 10
          default: 0
          description: Queue priority override.
      required: [source]
    TaskCreatedResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        message:
          type: string
          example: Document upload accepted. Processing started in background.
        websocket_url:
          type: string
          example: /ws/task/06d4ec78-20b2-4ac4-8c42-5dd52bcb10a6/progress
      required: [task_id, document_id, status, message, websocket_url]
    EvidenceItemSummary:
      type: object
      properties:
        id:
          type: string
        acmg_code:
          type: string
          description: ACMG evidence code (PS, PM, etc.).
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        review_required:
          type: boolean
        source_page:
          type: integer
          minimum: 1
    ParsingTaskStatusResponse:
      type: object
      properties:
        task_id:
          type: string
        document_id:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
        current_stage:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        evidence_items:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceItemSummary'
        processing_time_seconds:
          type: number
          format: float
          nullable: true
        file_size_bytes:
          type: integer
          nullable: true
      required:
        - task_id
        - document_id
        - status
        - progress_percentage
        - created_at
        - updated_at
    TaskProgressResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
        current_stage:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
      required: [task_id, status, progress_percentage, updated_at]
    TaskActionResponse:
      type: object
      properties:
        task_id:
          type: string
        message:
          type: string
      required: [task_id, message]
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true